Add files via upload
This version changes the decoding algorithm, so it sorts the decoded generation before doing row operations. This is done, because doing row operations without ordering the decoded generation is sometimes getting rid of independent coded symbols (The unknown bug that the algorithm still works with). Therefore, that bug is no longer unknown and is now fixed.
This version also separates the generation sorting function and the reduced low echelon form function, as they where all under one function before. This is because the decoding algorithm now uses the sorting function.
This version also adds a lot of print statements, so that the coded symbol can be followed through its journey in the decoding algorithm.
This version still needs to be tested. All the print statements might crash the LoPy ¯\_(ツ)_/¯