Update LoPy_Decoder_File_V3.py
Fixed a bug that made it not able to decode files with more than one generation. The fix was to make the Sort_Packet function make a check to see if the generation the coded packet belongs to is already decoded. Changed some stuff in the main function to be able to work with the new implementation (Generation decoded was in the main function before, it is now moved to Sort_Packet).